requirements.txt
numpy==1.24.2
tqdm==4.65.0
pytorch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
dgl==1.1
ogb==1.3.6
torchmetrics
scikit-optimize
matplotlib